BACKEND/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
python-dotenv
requests
Jinja2
firebase-admin
google-generativeai
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
email-validator
